Split the code
